requirements.txt
websockets>=12.0
firebase-admin>=6.5.0
python-dotenv>=1.0.0
requests>=2.31.0

